requirements.txt
tensorflow==2.18.0
streamlit==1.25.0
pillow==9.4.0
numpy==1.24.0